refactor(codebase): header guards and clean up comments across multiple files
- Removed unnecessary comments in header files and replaced them with standard header guards.
- Cleaned up comments in source files to improve readability and maintainability.
- Updated the `verify_admin_login` function to use `size_t` for index calculations.
- Streamlined the `create_bundle` and `remove_product_from_bundle` functions by removing redundant comments.
- Enhanced the `load_database` function to improve clarity by removing commented-out sections.
- Simplified test cases by removing excessive comments and focusing on essential assertions.

+++ src/admin.c
@@ -82,7 +82,6 @@ int verify_admin_login(const char *username, const char *password) {
     return 0;
   }
 
-  /* Obfuscate the input password to compare directly */
   char input_hex[(MAX_PASSWORD_LEN * 2) + 1] = {0};
   int input_len = (int)strlen(password);
   if (input_len >= MAX_PASSWORD_LEN) {
@@ -91,7 +90,7 @@ int verify_admin_login(const char *username, const char *password) {
   for (int i = 0; i < input_len; i++) {
     unsigned char c = (unsigned char)password[i];
     unsigned char obfuscated = (unsigned char)(c ^ 0x5A);
-    sprintf(input_hex + (i * 2), "%02X", obfuscated);
+    sprintf(input_hex + (size_t)(i * 2), "%02X", obfuscated);
   }
 
   if (strcmp(stored_hex, input_hex) == 0) {
